refactor(SSE): 패키지 이동
- 알림 외 서비스에서 이용 가능
