--- src/main/java/com/somemore/notification/sse/usecase/SseSender.java
@@ -1,7 +1,0 @@
-package com.somemore.notification.sse.usecase;
-
-import com.somemore.notification.sse.domain.Event;
-
-public interface SseSender {
-    <T> void send(Event<T> event);
-}
