+++ src/main/java/com/somemore/sse/subscriber/SseSubscriberImpl.java
@@ -1,7 +1,6 @@
-package com.somemore.notification.sse.service;
+package com.somemore.sse.subscriber;
 
-import com.somemore.notification.repository.emitter.EmitterRepository;
-import com.somemore.notification.sse.usecase.SseSubscriber;
+import com.somemore.sse.repository.emitter.EmitterRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 import org.springframework.transaction.annotation.Transactional;
